jwt implementation on lecture

+++ backend/controllers/lectureController.js
@@ -1,7 +1,10 @@
 import Lecture from "../models/LectureModel.js";
-import { createLectureValidation } from "../validation.js";
+import { createLectureValidation } from "../helpers/validation.js";
 
 export const create = async (req, res) => {
+	// Verify User if the user is Admin or not
+	if (!req.user.isSuperuser) return res.status(403).json({ message: "You're Prohibited to access!" });
+
 	// Validate the request
 	const { error } = createLectureValidation(req.body);
 	if (error) return res.status(401).json({ message: error.details[0].message });
@@ -16,12 +19,16 @@ export const create = async (req, res) => {
 		tags: req.body.tags,
 	});
 
-	try {
-		const createdLecture = await lecture.save();
-		res.status(201).json({ message: "Lecture Created!", data: createdLecture });
-	} catch (err) {
-		res.status(404).json({ message: err.message });
-	}
+	lecture
+		.save()
+		.then((result) => {
+			res.status(201).json({ message: "Lecture Created!", data: result });
+		})
+		.catch((err) => {
+			res.status(409).json({
+				message: err.message || "Some error while creating data!",
+			});
+		});
 };
 
 export const findAll = async (req, res) => {
@@ -73,19 +80,22 @@ export const update = async (req, res) => {
 };
 
 export const remove = async (req, res) => {
+	// Verify User if the user is Admin or not
+	if (!req.user.isSuperuser) return res.status(403).json({ message: "You're Prohibited to access!" });
+
 	const id = req.params.id;
 
 	Lecture.findByIdAndDelete(id)
 		.then((result) => {
-            if (!result) return res.status(404).json({ message: "Lecture not found..." });
+			if (!result) return res.status(404).json({ message: "Lecture not found..." });
 
-            res.json({
+			res.json({
 				message: "Lecture was deleted",
 			});
-        })
+		})
 		.catch((err) => {
-            res.status(409).json({
+			res.status(409).json({
 				message: err.message || "Some error while deleting lecture!",
 			});
-        });
+		});
 };
